refactor: host RootsOfUnity in op-service/kzg, off op-program
Move the EIP-4844 RootsOfUnity table out of op-program/client/l1 into a new
op-service/kzg package, so op-challenger no longer imports op-program.

op-service (rather than op-challenger) is the right home: op-program/client is
the fault-proof program compiled for cannon, and its build context (see
op-program/Dockerfile.vmcompat.dockerignore / Dockerfile.repro.dockerignore)
allowlists op-service but not op-challenger. Since client/l1 itself uses
RootsOfUnity, it must stay importable from the FPVM client; op-service is the
neutral package both op-challenger and op-program may depend on.

Prerequisite for deleting op-program from the monorepo (ethereum-optimism/optimism#20276).
