+++ op-challenger/game/fault/trace/utils/preimage_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
 	"github.com/ethereum-optimism/optimism/op-challenger/kvstore"
 	preimage "github.com/ethereum-optimism/optimism/op-preimage"
-	"github.com/ethereum-optimism/optimism/op-program/client/l1"
+	"github.com/ethereum-optimism/optimism/op-service/kzg"
 	"github.com/ethereum-optimism/optimism/op-service/testutils"
 )
 
@@ -80,7 +80,7 @@ func TestPreimageLoader_BlobPreimage(t *testing.T) {
 	indices := []uint64{0, 1, 24, 2222, 4095}
 	for _, fieldIndex := range indices {
 		elementData := blob[fieldIndex<<5 : (fieldIndex+1)<<5]
-		zPoint := l1.RootsOfUnity[fieldIndex].Bytes()
+		zPoint := kzg.RootsOfUnity[fieldIndex].Bytes()
 		kzgProof, claim, err := kzg4844.ComputeProof(&blob, zPoint)
 		require.NoError(t, err)
 		elementDataWithLengthPrefix := make([]byte, len(elementData)+lengthPrefixSize)
@@ -215,7 +215,7 @@ func storeBlob(t *testing.T, kv kvstore.KV, commitment kzg4844.Commitment, blob 
 	blobKeyBuf := make([]byte, 80)
 	copy(blobKeyBuf[:48], commitment[:])
 	for i := 0; i < params.BlobTxFieldElementsPerBlob; i++ {
-		root := l1.RootsOfUnity[i].Bytes()
+		root := kzg.RootsOfUnity[i].Bytes()
 		copy(blobKeyBuf[48:], root[:])
 		feKey := crypto.Keccak256Hash(blobKeyBuf)
 		err := kv.Put(preimage.Keccak256Key(feKey).PreimageKey(), blobKeyBuf)
