+++ op-service/kzg/roots_of_unity_test.go
@@ -1,0 +1,47 @@
+package kzg
+
+import (
+	"math/rand"
+	"testing"
+
+	"github.com/ethereum/go-ethereum/crypto/kzg4844"
+	"github.com/ethereum/go-ethereum/params"
+	"github.com/stretchr/testify/require"
+
+	"github.com/ethereum-optimism/optimism/op-service/testutils"
+)
+
+// TestInitRootsOfUnity validates that the roots of unity are constructed and ordered correctly such that the
+// root at index i can be used to compute the field element at index i in a blob
+func TestInitRootsOfUnity(t *testing.T) {
+	// Check we have the right number of roots
+	require.Equal(t, params.BlobTxFieldElementsPerBlob, len(RootsOfUnity))
+
+	// Create a blob with random data
+	rng := rand.New(rand.NewSource(123))
+	blob, blobCommitment, err := testutils.RandomBlob(rng)
+	require.NoError(t, err)
+
+	// Verify we can generate each field element using the ordered roots of unity
+	for i := 0; i < params.BlobTxFieldElementsPerBlob; i++ {
+		if i%100 == 0 {
+			t.Logf("Checking field element %d", i)
+		}
+
+		// Extract the target field element
+		var targetFieldElement [32]byte
+		copy(targetFieldElement[:], blob[i*32:(i+1)*32])
+
+		// We'll use the ith root of unity to evaluate the blob polynomial
+		z := RootsOfUnity[i].Bytes()
+
+		// Check that the correct field element is generated by evaluating the polynomial at point z
+		kzgProof, evaluation, err := kzg4844.ComputeProof(&blob, z)
+		require.NoError(t, err)
+		require.Equal(t, targetFieldElement[:], evaluation[:])
+
+		// Verify the proof for good measure
+		err = kzg4844.VerifyProof(blobCommitment, z, targetFieldElement, kzgProof)
+		require.NoError(t, err)
+	}
+}
